# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Vantage Download Site API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Link {
  id: string;
  uid: string;
  originalUrl: string;
  directUrl: string;
  shortCode: string;
  fileName: string;
  fileType: string;
  /** @nullable */
  fileSize?: string | null;
  downloads: number;
  createdAt: string;
  reported?: boolean;
}

export interface LinkInput {
  uid: string;
  originalUrl: string;
  /** @nullable */
  customSlug?: string | null;
}

export interface DownloadEvent {
  device: string;
  /** @nullable */
  country?: string | null;
}

export interface ReportInput {
  reason: string;
}

export interface User {
  uid: string;
  email: string;
  /** @nullable */
  displayName?: string | null;
  /** @nullable */
  photoURL?: string | null;
  role: string;
  linksUsed: number;
  linksLimit: number;
  tasksCompleted: string[];
  createdAt: string;
}

export interface UserUpdate {
  displayName?: string;
  photoURL?: string;
}

export interface TaskCompletion {
  task: string;
}

export interface RoleUpdate {
  role: string;
}

export type AnalyticsSummaryDeviceBreakdown = {[key: string]: number};

export interface AnalyticsSummary {
  totalLinks: number;
  totalDownloads: number;
  totalUsers: number;
  todayDownloads: number;
  todayVisitors: number;
  deviceBreakdown?: AnalyticsSummaryDeviceBreakdown;
  topFiles?: Link[];
}

export interface SiteSettings {
  socialTaskEnabled: boolean;
  directDownloadEnabled: boolean;
  monetizationEnabled: boolean;
  youtubeLink: string;
  instagramLink: string;
  whatsappLink: string;
  /** @nullable */
  adsensePublisherId?: string | null;
}

export interface SiteSettingsUpdate {
  socialTaskEnabled?: boolean;
  directDownloadEnabled?: boolean;
  monetizationEnabled?: boolean;
  youtubeLink?: string;
  instagramLink?: string;
  whatsappLink?: string;
  adsensePublisherId?: string;
}

export type GetLinksParams = {
uid: string;
};