# lib/api-zod/src/generated/types/analyticsSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Vantage Download Site API
 * OpenAPI spec version: 0.1.0
 */
import type { AnalyticsSummaryDeviceBreakdown } from './analyticsSummaryDeviceBreakdown';
import type { Link } from './link';

export interface AnalyticsSummary {
  totalLinks: number;
  totalDownloads: number;
  totalUsers: number;
  todayDownloads: number;
  todayVisitors: number;
  deviceBreakdown?: AnalyticsSummaryDeviceBreakdown;
  topFiles?: Link[];
}