# lib/api-zod/src/generated/types/link.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Vantage Download Site API
 * OpenAPI spec version: 0.1.0
 */

export interface Link {
  id: string;
  uid: string;
  originalUrl: string;
  directUrl: string;
  shortCode: string;
  fileName: string;
  fileType: string;
  /** @nullable */
  fileSize?: string | null;
  downloads: number;
  createdAt: string;
  reported?: boolean;
}