# lib/api-zod/src/generated/types/linkInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Vantage Download Site API
 * OpenAPI spec version: 0.1.0
 */

export interface LinkInput {
  uid: string;
  originalUrl: string;
  /** @nullable */
  customSlug?: string | null;
}