# lib/api-zod/src/generated/types/siteSettings.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Vantage Download Site API
 * OpenAPI spec version: 0.1.0
 */

export interface SiteSettings {
  socialTaskEnabled: boolean;
  directDownloadEnabled: boolean;
  monetizationEnabled: boolean;
  youtubeLink: string;
  instagramLink: string;
  whatsappLink: string;
  /** @nullable */
  adsensePublisherId?: string | null;
}