# lib/api-zod/src/generated/types/siteSettingsUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Vantage Download Site API
 * OpenAPI spec version: 0.1.0
 */

export interface SiteSettingsUpdate {
  socialTaskEnabled?: boolean;
  directDownloadEnabled?: boolean;
  monetizationEnabled?: boolean;
  youtubeLink?: string;
  instagramLink?: string;
  whatsappLink?: string;
  adsensePublisherId?: string;
}