# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Vantage Download Site API
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  uid: string;
  email: string;
  /** @nullable */
  displayName?: string | null;
  /** @nullable */
  photoURL?: string | null;
  role: string;
  linksUsed: number;
  linksLimit: number;
  tasksCompleted: string[];
  createdAt: string;
}